pyproject.toml
[project]
name = "wiki-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "mcp>=1.26.0",
    "python-decouple>=3.8",
    "ruff>=0.15.2",
    "wikipedia>=1.4.0",
    "wikipedia-sections>=2.0.0",
]
